+++ static/js/src/desktop-statistics.js
@@ -1,18 +1,5 @@
 import dummyData from "./dummy-data";
-
-function debounce(func, wait, immediate) {
-  var timeout;
-  return function () {
-    var context = this,
-      args = arguments;
-    clearTimeout(timeout);
-    timeout = setTimeout(function () {
-      timeout = null;
-      if (!immediate) func.apply(context, args);
-    }, wait);
-    if (immediate && !timeout) func.apply(context, args);
-  };
-}
+import { debounce } from "./utils/debounce.js";
 
 function calcPercentage(dataset, datum) {
   var sum = d3.sum(dataset, function (d) {
