+++ static/js/src/imageBuilder.js
@@ -1,3 +1,5 @@
+import { debounce } from "./utils/debounce.js";
+
 window.renderImageBuilder = function () {
   // State management
   class StateArray extends Array {
@@ -350,22 +352,6 @@ window.renderImageBuilder = function () {
       }
     }
     return false;
-  }
-
-  function debounce(func, wait, immediate) {
-    var timeout;
-    return function () {
-      var context = this,
-        args = arguments;
-      var later = function () {
-        timeout = null;
-        if (!immediate) func.apply(context, args);
-      };
-      var callNow = immediate && !timeout;
-      clearTimeout(timeout);
-      timeout = setTimeout(later, wait);
-      if (callNow) func.apply(context, args);
-    };
   }
 
   render();
